+++ hesma/meta/models.py
@@ -1,4 +1,3 @@
-from bootstrap_modal_forms.forms import BSModalModelForm
 from django.db import models
 
 
@@ -13,12 +12,3 @@ class DOI(models.Model):
     class Meta:
         ordering = ["-created"]
         verbose_name = "DOI"
-
-
-class DOIForm(BSModalModelForm):
-    class Meta:
-        model = DOI
-        fields = ["doi"]
-        labels = {"doi": "DOI"}
-        help_texts = {"doi": "Enter a DOI to add to the database."}
-        error_messages = {"doi": {"unique": "This DOI is already in the database."}}
